Various fixes and API renames
