+++ app/code/Magento/InventoryCatalog/etc/di.xml
@@ -8,10 +8,6 @@
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
     <preference for="Magento\InventoryCatalog\Api\DefaultStockProviderInterface" type="Magento\InventoryCatalog\Model\DefaultStockProvider"/>
     <preference for="Magento\InventoryCatalog\Api\DefaultSourceProviderInterface" type="Magento\InventoryCatalog\Model\DefaultSourceProvider"/>
-    <preference for="Magento\InventoryCatalog\Model\Command\UpdateCatalogInventoryStockItemByDefaultSourceItemInterface" type="Magento\InventoryCatalog\Model\Command\UpdateCatalogInventoryStockItemByDefaultSourceItem"/>
-    <preference for="Magento\InventoryCatalog\Model\Command\UpdateCatalogInventoryStockStatusByDefaultSourceItemInterface" type="Magento\InventoryCatalog\Model\Command\UpdateCatalogInventoryStockStatusByDefaultSourceItem"/>
-    <preference for="Magento\InventoryCatalog\Model\Command\DeleteCatalogInventoryStockItemByDefaultSourceItemInterface" type="Magento\InventoryCatalog\Model\Command\DeleteCatalogInventoryStockItemByDefaultSourceItem"/>
-    <preference for="Magento\InventoryCatalog\Model\Command\DeleteCatalogInventoryStockStatusByDefaultSourceItemInterface" type="Magento\InventoryCatalog\Model\Command\DeleteCatalogInventoryStockStatusByDefaultSourceItem"/>
     <type name="Magento\InventoryApi\Api\StockRepositoryInterface">
         <plugin name="prevent_default_stock_deleting"
                 type="Magento\InventoryCatalog\Plugin\InventoryApi\StockRepository\PreventDeleting\DefaultStockPlugin"/>
@@ -19,7 +15,10 @@
                 type="Magento\InventoryCatalog\Plugin\InventoryApi\StockRepository\PreventDeleting\AssignedToSalesChannelsStockPlugin"/>
     </type>
     <type name="Magento\InventoryApi\Api\SourceItemsSaveInterface">
-        <plugin name="update_legacy_catalog_inventory_at_source_items_save" type="Magento\InventoryCatalog\Plugin\InventoryApi\SaveStockItemsOnSourceItemsSavePlugin"/>
+        <plugin name="update_legacy_catalog_inventory_at_source_items_save" type="Magento\InventoryCatalog\Plugin\InventoryApi\UpdateLegacyCatalogInventoryOnSourceItemsSavePlugin"/>
+    </type>
+    <type name="Magento\InventoryApi\Api\SourceItemsDeleteInterface">
+        <plugin name="delete_legacy_catalog_inventory_plugin" type="Magento\InventoryCatalog\Plugin\InventoryApi\DeleteLegacyCatalogInventoryPlugin"/>
     </type>
     <type name="Magento\InventoryApi\Api\ReservationsAppendInterface">
         <plugin name="update_legacy_catalog_inventory_at_stock_deduction" type="Magento\InventoryCatalog\Plugin\InventoryApi\UpdateLegacyCatalogInventoryAtStockDeductionPlugin"/>
@@ -30,10 +29,10 @@
     <type name="Magento\CatalogInventory\Model\ResourceModel\Stock\Item">
         <plugin name="update_source_item_at_legacy_stock_setting"
                 type="Magento\InventoryCatalog\Plugin\CatalogInventory\UpdateSourceItemAtLegacyStockSettingPlugin"/>
+    </type>
+    <type name="Magento\CatalogInventory\Model\Stock\StockItemRepository">
         <plugin name="delete_source_items_when_legacy_stock_item_deleted"
-                type="Magento\InventoryCatalog\Plugin\CatalogInventory\Model\ResourceModel\Stock\DeleteSourceItemsAtLegacyStockItemDelete"/>
+                type="Magento\InventoryCatalog\Plugin\CatalogInventory\DeleteSourceItemsAtLegacyStockItemDelete"/>
     </type>
-    <type name="Magento\InventoryApi\Api\SourceItemsDeleteInterface">
-        <plugin name="delete_legacy_catalog_inventory_plugin" type="Magento\InventoryCatalog\Plugin\InventoryApi\DeleteLegacyCatalogInventoryPlugin"/>
-    </type>
+
 </config>
